login: Add 64-bit time support to utmp/utmpx
The new struct has the same size for 32-bit and 64-bit architecture with
two main differences compared to the 32-bit time one:

  1. ut_session is now 64-bit and it is mainly to set the ut_tv member
     alignment to be similar on both 32-bit and 64-bit architecture
     (for architectures that support multiple ABIs with different
     wordsize, such as x86 and powerpc).

  2. The internal tv_sec and tv_usec for ut_tv are now 64-bit.  Although
     it does not fully fix BZ#17470 nor BZ#18235, it might allow define
     the type as 'struct timeval' for __TIMESIZE=64 (thus fixing the
     aforementioned bug in this build configuration).

Different than laid out on the Y2038 Proofness Design [1], the
'utmp.trans' strategy is not used.  Instead:

  - New file names are defined for _PATH_UTMP, _PATH_WTMP, and
    _PATH_UTMP_DB (if applicable) and used as default for the new 2.34
    utmp{x} symbols.

  - The new utmp{x} symbols read/write 64-bit time_t entries as default.
    However if the old _PATH_UTMP or _PATH_WTMP is passed on
    utm{x}pname or updwtmp, the implementation read 32-bit utmp{x}
    entries and convert it to 64-bit ones.

  - The compat symbols read/write 32-bit time_t entries as default.
    If the default _PATH_UTMP or _PATH_WTMP is passed on
    utm{x}pname or updwtmp, the implementation reads 64-bit entries
    and convert to 32-bit ones.

The idea is not to maintain multiple databases with different formats
(which has underlying issues due to complexity, how to handle entries
that might overflow, and increases the security surface of BZ#24492),
but rather to move new application to use y2038 entries (which currently
affects 64-bit architecture as well, modulo s390).

If required the system might provide a tool to convert the old format
to newer one by opening the old file with utmpname and copying the
entries to the new format with updwtmp.

Also, for new 64-bit databases the path is not redirected to use the
utmpx one depending of the file existance (utmp_file_name_time32).
This is undocumented behavior most likely added be compatible with
Solaris (which defines different utmp dabases for utmpx files).

The s390 is an outlier: the 31-bit ABI added 64-bit time support on
GLIBC 2.8 and 64-bit has support since its inclusion.  The s390 ABI
follows the above design, but with a different ABI base version
(2.8 vs 2.34). The s390x instead does not have support to read/write
32-bit registers and does not provide compat symbols.

Checked on x86_64-linux-gnu and i686-linux-gnu.

[1] https://sourceware.org/glibc/wiki/Y2038ProofnessDesign#utmp_types_and_APIs